Move Wrapper classes to separate package
