+++ booking/src/main/java/com/drbsimon/booking/model/wrapper/ReservationWrapper.java
@@ -1,4 +1,4 @@
-package com.drbsimon.booking.model;
+package com.drbsimon.booking.model.wrapper;
 import com.drbsimon.booking.entity.Reservation;
 import lombok.AllArgsConstructor;
 import lombok.Data;
