+++ booking/src/main/java/com/drbsimon/booking/model/wrapper/ShowListWrapper.java
@@ -1,4 +1,4 @@
-package com.drbsimon.booking.model;
+package com.drbsimon.booking.model.wrapper;
 
 import com.drbsimon.booking.service.model.Show;
 import lombok.Data;
