+++ booking/src/main/java/com/drbsimon/booking/service/caller/VisitorServiceCaller.java
@@ -1,7 +1,7 @@
 package com.drbsimon.booking.service.caller;
 
 import com.drbsimon.booking.service.model.VisitorDTO;
-import com.drbsimon.booking.model.VisitorListWrapper;
+import com.drbsimon.booking.model.wrapper.VisitorListWrapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
